+++ pkg/frontend/nvme_controller.go
@@ -1,0 +1,274 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright (c) 2022-2023 Dell Inc, or its subsidiaries.
+// Copyright (C) 2023 Intel Corporation
+
+// Package frontend implememnts the FrontEnd APIs (host facing) of the storage Server
+package frontend
+
+import (
+	"context"
+	"fmt"
+	"log"
+	"net"
+	"path"
+	"sort"
+
+	"github.com/opiproject/gospdk/spdk"
+	pb "github.com/opiproject/opi-api/storage/v1alpha1/gen/go"
+	"github.com/opiproject/opi-spdk-bridge/pkg/server"
+
+	"github.com/google/uuid"
+	"go.einride.tech/aip/fieldbehavior"
+	"go.einride.tech/aip/fieldmask"
+	"go.einride.tech/aip/resourceid"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
+	"google.golang.org/protobuf/types/known/emptypb"
+)
+
+const (
+	ipv4NvmeTCPProtocol = "ipv4"
+	ipv6NvmeTCPProtocol = "ipv6"
+)
+
+// TODO: consider using https://pkg.go.dev/net#TCPAddr
+type tcpSubsystemListener struct {
+	listenAddr net.IP
+	listenPort string
+	protocol   string
+}
+
+func sortNvmeControllers(controllers []*pb.NvmeController) {
+	sort.Slice(controllers, func(i int, j int) bool {
+		return controllers[i].Name < controllers[j].Name
+	})
+}
+
+// NewTCPSubsystemListener creates a new instance of tcpSubsystemListener
+func NewTCPSubsystemListener(listenAddr string) SubsystemListener {
+	host, port, err := net.SplitHostPort(listenAddr)
+	if err != nil {
+		log.Panicf("Invalid ip:port tuple: %v", listenAddr)
+	}
+
+	parsedAddr := net.ParseIP(host)
+	if parsedAddr == nil {
+		log.Panicf("Invalid ip address: %v", host)
+	}
+
+	var protocol string
+	switch {
+	case parsedAddr.To4() != nil:
+		protocol = ipv4NvmeTCPProtocol
+	case parsedAddr.To16() != nil:
+		protocol = ipv6NvmeTCPProtocol
+	default:
+		log.Panicf("Not supported protocol for: %v", listenAddr)
+	}
+
+	return &tcpSubsystemListener{
+		listenAddr: parsedAddr,
+		listenPort: port,
+		protocol:   protocol,
+	}
+}
+
+func (c *tcpSubsystemListener) Params(_ *pb.NvmeController, nqn string) spdk.NvmfSubsystemAddListenerParams {
+	result := spdk.NvmfSubsystemAddListenerParams{}
+	result.Nqn = nqn
+	result.ListenAddress.Trtype = "tcp"
+	result.ListenAddress.Traddr = c.listenAddr.String()
+	result.ListenAddress.Trsvcid = c.listenPort
+	result.ListenAddress.Adrfam = c.protocol
+
+	return result
+}
+
+// CreateNvmeController creates an Nvme controller
+func (s *Server) CreateNvmeController(_ context.Context, in *pb.CreateNvmeControllerRequest) (*pb.NvmeController, error) {
+	log.Printf("Received from client: %v", in.NvmeController)
+	// check required fields
+	if err := fieldbehavior.ValidateRequiredFields(in); err != nil {
+		log.Printf("error: %v", err)
+		return nil, err
+	}
+	// check input parameters validity
+	if in.NvmeController.Spec == nil || in.NvmeController.Spec.SubsystemId == nil || in.NvmeController.Spec.SubsystemId.Value == "" {
+		return nil, status.Error(codes.InvalidArgument, "invalid input subsystem parameters")
+	}
+	// see https://google.aip.dev/133#user-specified-ids
+	resourceID := resourceid.NewSystemGenerated()
+	if in.NvmeControllerId != "" {
+		err := resourceid.ValidateUserSettable(in.NvmeControllerId)
+		if err != nil {
+			log.Printf("error: %v", err)
+			return nil, err
+		}
+		log.Printf("client provided the ID of a resource %v, ignoring the name field %v", in.NvmeControllerId, in.NvmeController.Name)
+		resourceID = in.NvmeControllerId
+	}
+	in.NvmeController.Name = server.ResourceIDToVolumeName(resourceID)
+	// idempotent API when called with same key, should return same object
+	controller, ok := s.Nvme.Controllers[in.NvmeController.Name]
+	if ok {
+		log.Printf("Already existing NvmeController with id %v", in.NvmeController.Name)
+		return controller, nil
+	}
+	// not found, so create a new one
+	subsys, ok := s.Nvme.Subsystems[in.NvmeController.Spec.SubsystemId.Value]
+	if !ok {
+		err := fmt.Errorf("unable to find subsystem %s", in.NvmeController.Spec.SubsystemId.Value)
+		log.Printf("error: %v", err)
+		return nil, err
+	}
+
+	params := s.Nvme.subsysListener.Params(in.NvmeController, subsys.Spec.Nqn)
+	var result spdk.NvmfSubsystemAddListenerResult
+	err := s.rpc.Call("nvmf_subsystem_add_listener", &params, &result)
+	if err != nil {
+		log.Printf("error: %v", err)
+		return nil, err
+	}
+	log.Printf("Received from SPDK: %v", result)
+	if !result {
+		msg := fmt.Sprintf("Could not create CTRL: %s", in.NvmeController.Name)
+		log.Print(msg)
+		return nil, status.Errorf(codes.InvalidArgument, msg)
+	}
+	response := server.ProtoClone(in.NvmeController)
+	response.Spec.NvmeControllerId = -1
+	response.Status = &pb.NvmeControllerStatus{Active: true}
+	s.Nvme.Controllers[in.NvmeController.Name] = response
+
+	return response, nil
+}
+
+// DeleteNvmeController deletes an Nvme controller
+func (s *Server) DeleteNvmeController(_ context.Context, in *pb.DeleteNvmeControllerRequest) (*emptypb.Empty, error) {
+	log.Printf("Received from client: %v", in.Name)
+	// check required fields
+	if err := fieldbehavior.ValidateRequiredFields(in); err != nil {
+		log.Printf("error: %v", err)
+		return nil, err
+	}
+	// fetch object from the database
+	controller, ok := s.Nvme.Controllers[in.Name]
+	if !ok {
+		if in.AllowMissing {
+			return &emptypb.Empty{}, nil
+		}
+		err := status.Errorf(codes.NotFound, "unable to find key %s", in.Name)
+		log.Printf("error: %v", err)
+		return nil, err
+	}
+	subsys, ok := s.Nvme.Subsystems[controller.Spec.SubsystemId.Value]
+	if !ok {
+		err := fmt.Errorf("unable to find subsystem %s", controller.Spec.SubsystemId.Value)
+		log.Printf("error: %v", err)
+		return nil, err
+	}
+
+	params := s.Nvme.subsysListener.Params(controller, subsys.Spec.Nqn)
+	var result spdk.NvmfSubsystemAddListenerResult
+	err := s.rpc.Call("nvmf_subsystem_remove_listener", &params, &result)
+	if err != nil {
+		log.Printf("error: %v", err)
+		return nil, err
+	}
+	log.Printf("Received from SPDK: %v", result)
+	if !result {
+		msg := fmt.Sprintf("Could not delete NQN:ID %s:%d", subsys.Spec.Nqn, controller.Spec.NvmeControllerId)
+		log.Print(msg)
+		return nil, status.Errorf(codes.InvalidArgument, msg)
+	}
+	delete(s.Nvme.Controllers, controller.Name)
+	return &emptypb.Empty{}, nil
+}
+
+// UpdateNvmeController updates an Nvme controller
+func (s *Server) UpdateNvmeController(_ context.Context, in *pb.UpdateNvmeControllerRequest) (*pb.NvmeController, error) {
+	log.Printf("UpdateNvmeController: Received from client: %v", in)
+	// check required fields
+	if err := fieldbehavior.ValidateRequiredFields(in); err != nil {
+		log.Printf("error: %v", err)
+		return nil, err
+	}
+	// fetch object from the database
+	volume, ok := s.Nvme.Controllers[in.NvmeController.Name]
+	if !ok {
+		if in.AllowMissing {
+			log.Printf("TODO: in case of AllowMissing, create a new resource, don;t return error")
+		}
+		err := status.Errorf(codes.NotFound, "unable to find key %s", in.NvmeController.Name)
+		log.Printf("error: %v", err)
+		return nil, err
+	}
+	resourceID := path.Base(volume.Name)
+	// update_mask = 2
+	if err := fieldmask.Validate(in.UpdateMask, in.NvmeController); err != nil {
+		log.Printf("error: %v", err)
+		return nil, err
+	}
+	log.Printf("TODO: use resourceID=%v", resourceID)
+	response := server.ProtoClone(in.NvmeController)
+	response.Status = &pb.NvmeControllerStatus{Active: true}
+	s.Nvme.Controllers[in.NvmeController.Name] = response
+	return response, nil
+}
+
+// ListNvmeControllers lists Nvme controllers
+func (s *Server) ListNvmeControllers(_ context.Context, in *pb.ListNvmeControllersRequest) (*pb.ListNvmeControllersResponse, error) {
+	log.Printf("Received from client: %v", in.Parent)
+	// check required fields
+	if err := fieldbehavior.ValidateRequiredFields(in); err != nil {
+		log.Printf("error: %v", err)
+		return nil, err
+	}
+	// fetch object from the database
+	Blobarray := []*pb.NvmeController{}
+	for _, controller := range s.Nvme.Controllers {
+		Blobarray = append(Blobarray, controller)
+	}
+	sortNvmeControllers(Blobarray)
+	token := uuid.New().String()
+	s.Pagination[token] = int(in.PageSize)
+	return &pb.ListNvmeControllersResponse{NvmeControllers: Blobarray, NextPageToken: token}, nil
+}
+
+// GetNvmeController gets an Nvme controller
+func (s *Server) GetNvmeController(_ context.Context, in *pb.GetNvmeControllerRequest) (*pb.NvmeController, error) {
+	log.Printf("Received from client: %v", in.Name)
+	// check required fields
+	if err := fieldbehavior.ValidateRequiredFields(in); err != nil {
+		log.Printf("error: %v", err)
+		return nil, err
+	}
+	// fetch object from the database
+	controller, ok := s.Nvme.Controllers[in.Name]
+	if !ok {
+		err := status.Errorf(codes.NotFound, "unable to find key %s", in.Name)
+		log.Printf("error: %v", err)
+		return nil, err
+	}
+	return &pb.NvmeController{Name: in.Name, Spec: &pb.NvmeControllerSpec{NvmeControllerId: controller.Spec.NvmeControllerId}, Status: &pb.NvmeControllerStatus{Active: true}}, nil
+}
+
+// NvmeControllerStats gets an Nvme controller stats
+func (s *Server) NvmeControllerStats(_ context.Context, in *pb.NvmeControllerStatsRequest) (*pb.NvmeControllerStatsResponse, error) {
+	log.Printf("NvmeControllerStats: Received from client: %v", in)
+	// check required fields
+	if err := fieldbehavior.ValidateRequiredFields(in); err != nil {
+		log.Printf("error: %v", err)
+		return nil, err
+	}
+	// fetch object from the database
+	volume, ok := s.Nvme.Controllers[in.Id.Value]
+	if !ok {
+		err := status.Errorf(codes.NotFound, "unable to find key %s", in.Id.Value)
+		log.Printf("error: %v", err)
+		return nil, err
+	}
+	resourceID := path.Base(volume.Name)
+	log.Printf("TODO: send anme to SPDK and get back stats: %v", resourceID)
+	return &pb.NvmeControllerStatsResponse{Stats: &pb.VolumeStats{ReadOpsCount: -1, WriteOpsCount: -1}}, nil
+}
